live-reload with periodic check (#5)
