+++ modules/translation/i18n/i18n_test.go
@@ -54,6 +54,9 @@ sub = Changed Sub String
 		langs, descs := ls.ListLangNameDesc()
 		assert.Equal(t, []string{"lang1", "lang2"}, langs)
 		assert.Equal(t, []string{"Lang1", "Lang2"}, descs)
-		_ = ls.Close()
+
+		result, found := ls.localeMap["lang1"].tryTr("no-such")
+		assert.Equal(t, "no-such", result)
+		assert.False(t, found)
 	}
 }
